requirements.txt
scikit-learn
numpy
pandas
matplotlib
scikit-learn
seaborn
statsmodels
google-generativeai
streamlit
moviepy
requests
os
re

